refactor(worker): isolate workspace result markers
