+++ workers/common/agent_runtime.go
@@ -8,7 +8,6 @@ package common
 
 import (
 	"context"
-	"crypto/rand"
 	"crypto/sha256"
 	"encoding/hex"
 	"encoding/json"
@@ -1438,139 +1437,6 @@ func runAgentInSandbox(ctx context.Context, name, workspaceDir string, sandboxEn
 	return nil
 }
 
-func forwardWorkspaceStdoutResultMarker(
-	ctx context.Context,
-	executor workspace.WorkspaceExecutor,
-	ref workspace.WorkspaceRef,
-	timeout time.Duration,
-	result *workspace.ExecResult,
-	expectedToken string,
-) error {
-	if !workerenv.IsTrue(os.Getenv(workerenv.ResultStdout)) {
-		return nil
-	}
-	marker, err := workspaceStdoutResultMarker(ctx, executor, ref, timeout, result, expectedToken)
-	if err != nil {
-		return err
-	}
-	if marker == "" {
-		return fmt.Errorf(
-			"%s is true but inner worker did not write %s",
-			workerenv.ResultStdout,
-			workerenv.ResultStdoutPrefix,
-		)
-	}
-	fmt.Println(marker)
-	return nil
-}
-
-func forwardWorkspaceStdoutResultMarkerIfPresent(
-	ctx context.Context,
-	executor workspace.WorkspaceExecutor,
-	ref workspace.WorkspaceRef,
-	timeout time.Duration,
-	result *workspace.ExecResult,
-	expectedToken string,
-) {
-	if !workerenv.IsTrue(os.Getenv(workerenv.ResultStdout)) {
-		return
-	}
-	marker, err := workspaceStdoutResultMarker(ctx, executor, ref, timeout, result, expectedToken)
-	if err != nil {
-		fmt.Fprintf(os.Stderr, "warning: failed to forward stdout result marker: %v\n", err)
-		return
-	}
-	if marker != "" {
-		fmt.Println(marker)
-	}
-}
-
-func workspaceStdoutResultMarker(
-	ctx context.Context,
-	executor workspace.WorkspaceExecutor,
-	ref workspace.WorkspaceRef,
-	timeout time.Duration,
-	result *workspace.ExecResult,
-	expectedToken string,
-) (string, error) {
-	if !workerenv.IsTrue(os.Getenv(workerenv.ResultStdout)) {
-		return "", nil
-	}
-
-	if result != nil {
-		if result.StdoutTruncated {
-			marker, downloadErr := downloadStdoutResultMarker(ctx, executor, ref, timeout, expectedToken)
-			if marker != "" {
-				return marker, nil
-			}
-			if downloadErr != nil {
-				return "", fmt.Errorf("download stdout result marker after truncated stdout: %w", downloadErr)
-			}
-			return "", fmt.Errorf(
-				"inner worker stdout was truncated before %s could be forwarded and marker file was not available",
-				workerenv.ResultStdoutPrefix,
-			)
-		}
-		if marker, ok := stdoutResultMarker(result.Stdout); ok {
-			return marker, nil
-		}
-	}
-	marker, downloadErr := downloadStdoutResultMarker(ctx, executor, ref, timeout, expectedToken)
-	if marker != "" {
-		return marker, nil
-	}
-	if downloadErr != nil && !workspace.IsKind(downloadErr, workspace.ErrorKindNotFound) {
-		return "", fmt.Errorf("download stdout result marker: %w", downloadErr)
-	}
-	return "", nil
-}
-
-func downloadStdoutResultMarker(
-	ctx context.Context,
-	executor workspace.WorkspaceExecutor,
-	ref workspace.WorkspaceRef,
-	timeout time.Duration,
-	expectedToken string,
-) (string, error) {
-	if executor == nil || ref.IsZero() {
-		return "", workspace.NewError(
-			"download",
-			workspace.ErrorKindNotFound,
-			"workspace reference is unavailable",
-			false,
-			nil,
-		)
-	}
-	result, err := executor.Download(ctx, workspace.DownloadRequest{
-		Ref:     ref,
-		Paths:   []string{agentSandboxResultMarkerUploadPath},
-		Timeout: timeout,
-	})
-	if err != nil {
-		return "", err
-	}
-	for _, artifact := range result.Artifacts {
-		if artifact.Path != agentSandboxResultMarkerUploadPath {
-			continue
-		}
-		data := string(artifact.Data)
-		if err := validateStdoutResultToken(data, expectedToken); err != nil {
-			return "", err
-		}
-		if marker, ok := stdoutResultMarker(data); ok {
-			return marker, nil
-		}
-		return "", fmt.Errorf("downloaded stdout result marker did not contain %s", workerenv.ResultStdoutPrefix)
-	}
-	return "", workspace.NewError(
-		"download",
-		workspace.ErrorKindNotFound,
-		"stdout result marker artifact not found",
-		false,
-		nil,
-	)
-}
-
 func stdoutResultMarker(stdout string) (string, bool) {
 	var marker string
 	for line := range strings.SplitSeq(stdout, "\n") {
@@ -1603,6 +1469,14 @@ func sandboxExecMaxOutputBytes() int64 {
 		return agentSandboxStdoutResultMaxOutputBytes
 	}
 	return agentSandboxExecMaxOutputBytes
+}
+
+func generateWorkspaceStdoutResultToken() (string, error) {
+	var random [32]byte
+	if _, err := rand.Read(random[:]); err != nil {
+		return "", fmt.Errorf("generate stdout result token: %w", err)
+	}
+	return hex.EncodeToString(random[:]), nil
 }
 
 func executionWorkspaceExecutor(workspaceEnv workerenv.ExecutionWorkspaceEnv) (workspace.WorkspaceExecutor, error) {
@@ -1763,14 +1637,6 @@ func stageAgentSandboxExecutable(
 	}
 	command = append(command, args...)
 	return command, innerEnv, nil
-}
-
-func generateWorkspaceStdoutResultToken() (string, error) {
-	var random [32]byte
-	if _, err := rand.Read(random[:]); err != nil {
-		return "", fmt.Errorf("generate stdout result token: %w", err)
-	}
-	return hex.EncodeToString(random[:]), nil
 }
 
 func agentSandboxWorkerCommand(tokenUploaded, gitAskpassUploaded bool, tokenCleanupPaths ...string) string {
